Use delimiter based parsing rules for template tag attribute values
Whitespace now ends the value unless within a ', ", (), [] or {} delimiter. This avoids issues with following content getting pulled in as a function arg. e.g.
"div attr-1=@foo.func not-an-arg=5" used to get parsed as "(@foo.func not) - an - arg)" (the '=' terminating because the current script support does not include such assignments).

Closes #86